optimize spots-router
