+++ backend/db/models/user.js
@@ -17,7 +17,9 @@ module.exports = (sequelize, DataTypes) => {
       })
 
       User.hasMany(models.Review, {
-        foreignKey: 'userId'
+        foreignKey: 'userId',
+        onDelete: 'CASCADE', 
+        hooks: true
       })
     }
   };
