+++ backend/routes/api/reviews.js
@@ -1,0 +1,13 @@
+
+
+// Get all Reviews of the Current User
+
+// Get all Reviews by a Spot's id
+
+// Create a Review for a Spot based on the Spot's id
+
+// Add an Image to a Review based on the Review's id
+
+// Edit a Review
+
+// Delete a Review
